Fronted initial

+++ src/main/java/com/domain/BoardGame.java
@@ -36,7 +36,7 @@ public class BoardGame {
     private Order order;
 
 
-    public BoardGame(String title, Float price, int copies) {
+    public BoardGame(String title, double price, int copies) {
         this.title = title;
         this.price = price;
         this.copies = copies;
